+++ src/admin/messages/el/cmsadmin.php
@@ -258,4 +258,8 @@ return [
 
     'menu_group_item_env_websites' => 'Websites',
     'model_website_use_default_theme' => 'Use default theme',
+    'model_website_all' => 'All',
+    'model_website_group_ids_label' => 'Restrict to User Groups',
+    'model_website_user_ids_label' => 'Restrict to Users',
+    'model_website_access_restrict' => 'Access restriction',
 ];