+++ CHANGELOG.md
@@ -5,6 +5,7 @@ In order to read more about upgrading and BC breaks have a look at the [UPGRADE 
 
 ## 4.0.0
 
+* [#345](https://github.com/luyadev/luya-module-cms/pull/345) Added Website user and group permission support
 * [#344](https://github.com/luyadev/luya-module-cms/pull/344) Added website collapse to page permission.
 * [#274](https://github.com/luyadev/luya-module-cms/pull/274) Multiple website support.
 * [#341](https://github.com/luyadev/luya-module-cms/pull/341) All deprecated methods has been removed `Url::toMenuItem`, `Block::objectId`, `NavItemPage::getBlock`.